+++ frontend/src/containers/List.jsx
@@ -1,0 +1,7 @@
+import PostComponent from "../components/PostComponent";
+
+function List() {
+  return <PostComponent type={"lists"} />;
+}
+
+export default List;
